Extract the magic templates into their own header to clean up core-types.h
